Create a nested structure

This function results in a two column data frame with nested structure. Currently only one parent is supported and child is only specified by giving the number of levels. (This will change shortly).

nest_in(
  .x,
  ...,
  name = "child",
  name_parent = NULL,
  prefix = NULL,
  suffix = NULL,
  unique = FALSE,
  leading0 = FALSE
)

Arguments

.x
A vector where each entry is a parent.

...
A single integer or sequence of two-sided formula. If a single integer then each parent will have children specified by that integer. If it is sequence of two-sided formula, then the left hand side (LHS) specifies the level as an integer or character. E.g. 1 means the first unique entry of the parent vector. If it is a character then it is assumed that it corresponds to the label of the parental level. Vector is supported for LHS. The right hand side (RHS) only supports numbers at the moment and corresponds to the number of children for the parental levels specified on LHS of the corresponding formula.

name
The name of the child variable.

name_parent
The name of the parent variable. If the parent was parsed as an assigned object then the name of the object is taken as the name unless name_parent is specified.

prefix
The prefix for the child labels.

suffix
The suffix for the child labels.

unique
A logical value to indicate whether the child labels across parents should be unique.

leading0
By default it is FALSE. If TRUE, this is the same as setting 0 or 1. If a positive integer is specified then it corresponds to the minimum number of digits for the child labels and there will be leading zeros augmented so that the minimum number is met.

Value

A two column data frame with the first column corresponding to parental levels and the second column corresponding to the child levels.

Examples

# Each element in the supplied the vector has 4 child.
nest_in(1:3, 4)

#>    parent child
#> 1       1     1
#> 2       1     2
#> 3       1     3
#> 4       1     4
#> 5       2     1
#> 6       2     2
#> 7       2     3
#> 8       2     4
#> 9       3     1
#> 10      3     2
#> 11      3     3
#> 12      3     4

# if an object pointing to the vector is supplied then the
# name of the object is used as a column name instead
first_name <- c("Tom", "Helen")
nest_in(first_name, 4)

#>   first_name child
#> 1        Tom     1
#> 2        Tom     2
#> 3        Tom     3
#> 4        Tom     4
#> 5      Helen     1
#> 6      Helen     2
#> 7      Helen     3
#> 8      Helen     4

# the variable name for parent and child can be overwritten
nest_in(first_name, 4, name = "pet", name_parent = "person")

#>   person pet
#> 1    Tom   1
#> 2    Tom   2
#> 3    Tom   3
#> 4    Tom   4
#> 5  Helen   1
#> 6  Helen   2
#> 7  Helen   3
#> 8  Helen   4

# prefix and suffix can be added to child labels
# along with other aesthesitics like leading zeroes
# with minimum number of digits.
nest_in(1:3, 10, prefix = "id-", suffix = "xy", leading0 = 4)

#>    parent     child
#> 1       1 id-0001xy
#> 2       1 id-0002xy
#> 3       1 id-0003xy
#> 4       1 id-0004xy
#> 5       1 id-0005xy
#> 6       1 id-0006xy
#> 7       1 id-0007xy
#> 8       1 id-0008xy
#> 9       1 id-0009xy
#> 10      1 id-0010xy
#> 11      2 id-0001xy
#> 12      2 id-0002xy
#> 13      2 id-0003xy
#> 14      2 id-0004xy
#> 15      2 id-0005xy
#> 16      2 id-0006xy
#> 17      2 id-0007xy
#> 18      2 id-0008xy
#> 19      2 id-0009xy
#> 20      2 id-0010xy
#> 21      3 id-0001xy
#> 22      3 id-0002xy
#> 23      3 id-0003xy
#> 24      3 id-0004xy
#> 25      3 id-0005xy
#> 26      3 id-0006xy
#> 27      3 id-0007xy
#> 28      3 id-0008xy
#> 29      3 id-0009xy
#> 30      3 id-0010xy

# you can specify unbalanced nested structures
nest_in(2:4,
         1 ~ 3,
         2 ~ 4,
         3 ~ 2)

#> Error in rep(.x, times = reps[.x]): invalid 'times' argument

# A `.` may be used to specify "otherwise".
nest_in(c("A", "B", "C", "D"),
             2:3 ~ 10,
               . ~ 3)

#>    parent child
#> 1       A     1
#> 2       A     2
#> 3       A     3
#> 4       B     1
#> 5       B     2
#> 6       B     3
#> 7       B     4
#> 8       B     5
#> 9       B     6
#> 10      B     7
#> 11      B     8
#> 12      B     9
#> 13      B    10
#> 14      C     1
#> 15      C     2
#> 16      C     3
#> 17      C     4
#> 18      C     5
#> 19      C     6
#> 20      C     7
#> 21      C     8
#> 22      C     9
#> 23      C    10
#> 24      D     1
#> 25      D     2
#> 26      D     3

# The parental level can be referred by its name or vectorised.
nest_in(c("A", "B", "C"),
         c("A", "B") ~ 10,
                 "C" ~ 3)

#>    parent child
#> 1       A     1
#> 2       A     2
#> 3       A     3
#> 4       A     4
#> 5       A     5
#> 6       A     6
#> 7       A     7
#> 8       A     8
#> 9       A     9
#> 10      A    10
#> 11      B     1
#> 12      B     2
#> 13      B     3
#> 14      B     4
#> 15      B     5
#> 16      B     6
#> 17      B     7
#> 18      B     8
#> 19      B     9
#> 20      B    10
#> 21      C     1
#> 22      C     2
#> 23      C     3